[PATCH] assign modified controller object in caller's workspace
functions which operate over controller objects in such a way that
the object itself is being updated, an assignment in caller's workspace
is made, such that it is possible to use these functions without any
output arguments. for instance

mpt_simplify(ctrl)

will simplify the controller stored in variable "ctrl" and automatically
update the object also in the parrent function. it is therefore
equivalent to calling

ctrl = mpt_simplify(ctrl)