[PATCH] assign modified polytope object in caller's workspace
in order to mitigate computational overhead, extreme points of a
polytope are stored in the polytope object itself. in order to increase
user's comfort, a given polytope object is automatically updated
in the caller's workspace, i.e. when one does

E = extreme(P)

the extreme points "E" are automatically stored in the polytope
object "P".

NOTE! NOTE! NOTE!
this automatic assignment does not work if the function is called
on an indexed polyarray, i.e. in case of

E = extreme(P(1))

"P(1)" will NOT be updated automatically with extreme points. in this
case it is necessary to do the assignment manually:

[E, R, P(1)] = extreme(P(1))