[PATH] save mptOptions permanently using setpref()
utilizes Matlab's setpref() and getpref() functions to store global
global settings permanently to disk. We utilize this feature to avoid
checking for available solvers at every mpt_init() call. Once the
settings are stored, they are read ONLY IF mpt_init() is called with
no input arguments. If some arguments are provided, check for
available solvers is always triggered.

NOTE! once the setting is stored in this way and the user removes
some solver(s), mpt_init will still think the removed solver is still
available. This may lead to crashes in cases where a LP/QP is
re-solved with a different solver if the main solver fails. In order to
rehash the solvers, one has to call:

mpt_init('rehash')