[PATCH] precompute matrices of MPC problem for MLD systems
Signed-off-by: Michal Kvasnica

constructs matrices of the MPC problem for MLD systems and stores
them into ctrl.details.Matrices (similar to what we do for LTI systems).
this should result in significant speedup in controller evaluation, since
we don't have to construct the problem all the time.

NOTE! if the controller is a tracking controller, matrices are not being
constructed!