[PATCH] reject systems with uncertainties for tracking
Signed-off-by: Michal Kvasnica

TODO: we can lift this limitation once we support flat noise polytopes.
the problem here is that noise polytope has to be augmented, because
the state vector included new states (past input and/or reference).

parametric uncertainties are easy to handle, but we postpone it and
deal with them when we also handle additive uncertainties.