[PATCH] calculate correct output for PWA/MLD systems
Signed-off-by: Michal Kvasnica

Spotted by Peter Ortner, this commit fixes issue138:

The output y is always calculated in the function "mdlOutputs" and if 
there is no direct feedtrough (D-Matrix of the LTI system is zero) then
the u-vector is set to zero! and the problem is now, that you first 
have to choose the valid PWA region before setting u to zero because 
the selection of the valid region not only depends on the actual state 
(guardX) it also depends on the actual input (guardU). When setting the
input to zero you get a wrong region.