[PATCH] solve LP with different solver if the current one fails
Signed-off-by: Michal Kvasnica

it was noticed that some LP solvers have difficulties when solving
the facet-LP. if we see problems with the default solver, we try
to solve the LP again with all available LP solvers until we find a
feasible solution. information about which solvers are being tried
is displayed if Options.verbose = 2

partially fixes issue138.