[PATCH] use unreduced H and K pair for checking correctness of hull
Signed-off-by: Mario Vasak
Signed-off-by: Michal Kvasnica

the command P=polytope(H,K,0,2) just could return an empty
polytope if cheby radius of P falls below the abs_tol, and that is 
correct. But taking H & K then from P brings empty H and K in the
superimposed function, and this ends in a hull-error (result=1), but
nothing is really wrong: H and K are hopefully correctly computed, but 
the resulting hull is too small, so H and K should be checked in 
checkhull.m, while hull.m should return an empty polytope if the check 
is all right.