[PATCH] if envelope is R^n, use the dimension of input polytope
Signed-off-by: Michal Kvasnica

before, R^n was returned as a polytope in R^1 with x <= Inf. this,
however, caused problems when performing operations which check
dimensions, as for instance the and (&) operator:

take for instance polyarray "P" in R^2, whose envelope is R^n. then
without this patch dimension(envelope(P)) would return 1. with this
patch applied, it returns 2 (or whatever the dimension of P is).