[PATCH] don't check regions twice
before we used to call isinside() on the whole controller partition
twice: first time with Options.fastbreak=1 to find out if the current
state belongs at least to the feasible set, and the second time we
called it with Options.fastbreak=0 to identify the proper feedback
law.

this redundancy was causing quite dramatic overhead in on-line
implementation of control laws.

now we called isinside() just once, with Options.fastbreak=0. this
brings a speedup of at least 50% even with simplest partitions.