speed improvements and bug fix
1. bug fix - In Pn(1) dynamics is such that you totally enter Pn(2) in 
one step, and in Pn(2) it is such that you leave the whole Pn in one 
step. Thus, the invariant set is an empty set. The old version returned
Pn(1) as the invariant set, which is wrong. New version now correctly
returns an empty set

2. speed improvements - implement handling of target partitions
according to "Rakovic, Grieder, Kvasnica, Mayne, Morari: Computation of
Invariant Sets for Piecewise Affine Discrete Time Systems subject to
Bounded Disturbances."