polytope/bounding_box(): change behavior of Options.noPolyOutput
if Options.noPolyOutput is set to true, we have been returning an
empty matrix as the first output argument, instead of returning
the full bounding box polytope (which is expensive to compute).

now with Options.noPolyOutput=1 we return the input polytope as
the first output argument, but we update the bounding box
information inside of that object, if it was not there.

example:
P = polytope(...);
P = set(P, 'bbox' []);  % clear bounding box information
Options.noPolyOutput=1;
[B, low, up] = bounding_box(P)

now B will be equal to P, but it will contain the bounding box
in B.bbox. Polytope arrays work in the same way.

unit test added: polytope/bounding_box_test3.m