mpt_plotPWQ, mpt_plotPWA: improved handling of new plots / holded plots
when mpt_plotPWQ() or mpt_plotPWA() are used several times, they would
plot the result into the same window on top of the previous result,
regardless of the status of 'hold on'.

now we always clear the current (sub)plot, but respect the user's
choice of 'hold on' or 'subplot()'.

this patch also simplifies the messy code of mpt_plotPWQ:
1. manual computation of bounding boxes replaced by calling
   bounding_box() directly (note that it supports polytope arrays)

2. if there are overlapping regions, for a given point we now plot
   the minimal function value associated to that point