mpt_yalmipcftoc: prefer MLD models for on-line controllers
from a large number of examples it seems that MLD models are
more efficient to solve compared to the approach where we
use one binary per region of a PWA model.

therefore we now use the MLD representation by default for
on-line controllers and give the user to force either PWA
or MLD models by setting Options.force_pwa or Options.force_mld
to 1.