mpt_yalmipcftoc: set bounds on guards for piecewise nonlinear systems
if system is piecewise nonlinear, we are using the implies() operator.
this, however, gives a warning if the expression for guards is
not bounded. currently YALMIP does not autopropagate bounds in
"constraint" objects (it would be very difficult to do so, see
comments in the patch).

therefore to avoid the warnings we just set large bounds on the
guards expressions. yes, this will probably lead to very poor
big-M relaxation, but there is no clear better (and simple) way of
doing it.