mpt_getInput: use the "global" solver for nonlinear problems by default
According to Johan, it's better to always use the "global" nonlinear
solver instead of the "local" version. However, it was observed on
numerous cases that if fmincon() is used as the global solver, it
very often just stalls under Matlab R13. No problems however when
Matlab R14 is used. It is therefore recommended to always use
Options.nlsolver='local' and Options.convertconvex=1 under Matlab R13.