mpt_getInput: work around problems with fmincon on Matlab R13
fmincon in R13 sets the MaxSQPIter option to Inf, which very causes
the solver to freeze. Matlab R14 uses a much better bound which is
a function of the number of optimization variables and constraints.
Therefore to prevent the cycling in R13/fmincon we now use the same
settings as R14 does. If problems remain, one can set the option
by changing Options.MaxSQPIter.