mpt_getInput: use smaller default value of Options.MaxSQPIter
Value of Options.MaxSQPIter is one of the most important factors which
influence convergence of the fmincon() solver. It was observed on the
nonlin1.m demo that the default value was too high, which caused
fmincon() to hang under R13. Therefore now we set a better upper
bound on this option, calculated from the number of variables and
constraints present in the problem.

This patch also adds support for Options.lowersolver, which is
propagated into sdpsettings('bmibnb.lowersolver', Options.lowersolver).
For instance the nonline1.m demo does not currently work with NAG
as a lower solver, but works fine with GLPK/CDD.