mpt_yalmipcftoc: augment x0 for tracking also for on-line controllers
if we have a system which we can augment to cope with tracking and/or
deltaU formulation (i.e. LTI or PWA systems), we now do so also for
on-line controllers (as we did in MPT 2.5). the advantage being
that on hand the optimization problem is consistent with its
"explicit" counterpart; moreover this allows us to solve on-line
problems which minimize deltaU instead of just u(k).

related to issue216