mpt_yalmipcftoc: return more info in case of on-line controllers
when constructing an on-line MPC problem, by changeset c86f3d49b238
we now augment the state vector to deal with tracking/deltaU
constraints whenever possible. in order to indicate that the reference
and/or the value of previous control action are now part of the state
vector (and not given as a separate variables), we now return this
in a special structure (1st output from mpt_yalmipcftoc(), unused
unless Options.dont_solve=1 is given). this additional information is
then processed by mptctrl().