mptctrl: improve handling of on-line controllers with tracking
by changeset c86f3d49b238 we now augment the state vector to cope
with tracking/deltaU formulation whenever possible. this implies that
we can handle probStruct.tracking=1 for LTI/PWA systems and we switch
to probStruct.tracking=2 for MLD/nonlinear systems.

in MPT 2.6 we always switched to tracking=2, which introduced an
inconsistency with MPT 2.5. therefore we now respect the tracking
flag as set by the user and use tracking=1 if possible.

this implies that reference/u(k-1) are part of the state vector if
tracking=1 is used (it is also true for tracking=2 with LTI/PWA
systems). this patch serves to recognize that the state vector was
augmented by exploting additional information provided by
mpt_yalmipcftoc() (changeset 63210f246295).