improve simulation of on-line MPC controllers
some on-line controllers now use an augmented state vector which
includes the reference and/or the previous control action. therefore
this patch is needed to adapt the simulation routines to this behavior.
namely, we need to compute a "true" control input in case where
the optimization used the deltaU formulation.

related to changesets 97ca6f0f0d8b and c86f3d49b238