mpt_yalmipcftoc: switch from tracking=1 to tracking=2 for MLD systems
if probStruct.tracking=1 is used together with MLD/nonlinear systems,
we now automatically switch to tracking=2 and display a message that
we have done so.

the reason being that tracking=1 implies deltaU formulation for which
we can't go in case of MLD/nonlinear dynamics. only tracking=2
formulations are allowed.

notice that such automatic switch will only happen for on-line
controllers, since explicit ones do require PWA/LTI models.