mptctrl: improve usage of remove_equalities
the easiest way how to decide whether or not we should remove
equalities is to look whether we have integer variables in the
model. if not, we can remove equality constraints to improve
performance.

notice that this way of handling is important if the problem
comes from mpt_ownmpc() - there we don't know whether the user
used a linear system.