mptctrl: keep information about the format of x0
ctrl.details.x0format is now a structure which defines 3 things:
1. dimension of the state vector
2. number of inputs stored in the state vector
3. number of references stored in the state vector

this information is important if we augmented the state vector
to deal with tracking/deltaU formulation as is later used in
the extendx0() helper.

the reason why we keep such info directly in the controller object
is to reduce the effort of simulation functions which would
otherwise need to perform a tricky analysis of the controller
object in order to find whether:
1. we minimize deltaU=u(k)-u(k-1) instead of u(k)
2. we need to extend x0 by adding u(k-1) and the reference

this patch also introduces a "loadobj" method which is needed to
assign "ctrl.details.x0format" also for objects saved with older
versions.