mpt_yalmipcftoc: use set() instead of bounds()
YALMIP released beyond 15-May-2006 deprecates the bounds() operator and
extracts bounds on variables directly from constraints. therefore we
have to add bounds as constraints in our models.

notice that since bounds are required when using the implies() operator
(which happens if we have a PWA or a piecewise nonlinear model), we now
force bounds of +/- mptOptions.infbox on states/inputs/outputs even
if they are not specified by the user. we give a nice warning that
we have done so, though.