mpt_optInfControlPWA: use a valid initial controller structure
if no core was found, we fall back into a standard CFTOC approach
where we detect regions belonging to the core. once we find at least
such region, we restart the algorithm with Options.core. however, the
generate core was not a valid controller structure due to a missing
"details" field.