bug 43: fprintf(0, ...) throws an error in R2006b
Commands such as fprintf(0, ...) and fwrite(0, ...), in which the file
identifier is zero (the same as stdin) in Matlab R2006b result in an
error being thrown. In previous releases, MATLAB did not throw an error
in response to these commands, even though printing or writing to stdin
is clearly not a valid option.