bug 47: suppress annoying warnings when using quadprog
If quadprog is used as a QP solver, it very often complains about a
non-symmetric Hessian, which is produced by an mpQP solver.

If the difference in symmetricity is only small, we should set the
Hessian to (H+H')*0.5 in mpt_solveQP to prevent these unnecessary
warnings.

unit test: solvers/mpt_solveQP_test5.m