Bug 58: Mex version of polytope/isinside.m
Since isinside() contributes heavily (by ~25%) to the total runtime
of mpqp solvers on large examples, it is desirable to have a C-coded
mex version of isinside().

For now we let the mex file uncompiled in the @polytope directory
and let users to compile it when speed is desired.

This version mirrors the functionality of polytope/isinside.m, with
the only exception of not allowing the call "isinside(P)", which
plots  the regions and lets the users to specify the point x0 by a
mouse click.

unit test: regress/polytope/isinside_test1.m
unit test: regress/polytope/isinside_test2.m