mpt_ownmpc: always change empty elements of ctrl.Ai to zeros
if the cost of the problem is not quadratic, YALMIP would return
the quadratic term of the cost (ctrl.Ai) set to an empty matrix. We
used to change such empty fields to zero matrices of appropriate
dimensions, but only if probStruct.norm was not set to 2. Using
mpt_ownmpc(), however, the user could manually change the cost
to zero, which will yield an mpLP, despite having probStruct.norm=2.

Therefore we now always change all empty elements of ctrl.Ai to
zeros.

unit test: regress/mpt_ownmpc_test20.m