sub_computeTrajectory: always pass known dynamics to mpt_simSys
if the index of the active PWA dynamics is known (e.g. from the respective
ctrl.dynamics field) then we pass this known index to mpt_simSys() to get
a correct state update.

note that this cal lead in discrepancies in the closed-loop behavior
when compared to simulink-based simulations. when the sim() and simplot()
functions are used, we let the controller decide which dynamics should be
used to calculate the state update. in Simulink there is no way how the
controller could "enforce" the active dynamics to the system. therefore
if a state lies on a boundary of two different PWA dynamics, different
choices can be made. in sub_computeTrajectory we always take the
"correct" dynamics (correct from the point of view of the controller).